Move code in retroarch.c over to configuration file
