+++ configuration.h
@@ -28,6 +28,7 @@
 #include "gfx/video_defines.h"
 #include "input/input_defines.h"
 #include "led/led_defines.h"
+#include "msg_hash.h"
 
 #ifdef HAVE_LAKKA
 #include "lakka.h"
@@ -63,6 +64,8 @@
    strlcpy(var, newvar, sizeof(var)); \
 }
 
+#define INPUT_CONFIG_BIND_MAP_GET(i) ((const struct input_bind_map*)&input_config_bind_map[(i)])
+
 enum crt_switch_type
 {
    CRT_SWITCH_NONE = 0,
@@ -81,6 +84,25 @@ enum override_type
 };
 
 RETRO_BEGIN_DECLS
+
+/* Input config. */
+struct input_bind_map
+{
+   const char *base;
+
+   enum msg_hash_enums desc;
+
+   /* Meta binds get input as prefix, not input_playerN".
+    * 0 = libretro related.
+    * 1 = Common hotkey.
+    * 2 = Uncommon/obscure hotkey.
+    */
+   uint8_t meta;
+
+   uint8_t retro_key;
+
+   bool valid;
+};
 
 typedef struct settings
 {
@@ -1048,6 +1070,27 @@ settings_t *config_get_ptr(void);
 const char *config_get_all_timezones(void);
 void config_set_timezone(char *timezone);
 #endif
+
+bool input_config_bind_map_get_valid(unsigned bind_index);
+
+void input_config_parse_joy_button(
+      char *s,
+      void *data, const char *prefix,
+      const char *btn, void *bind_data);
+
+void input_config_parse_joy_axis(
+      char *s,
+      void *conf_data, const char *prefix,
+      const char *axis, void *bind_data);
+
+void input_config_parse_mouse_button(
+      char *s,
+      void *conf_data, const char *prefix,
+      const char *btn, void *bind_data);
+
+const char *input_config_get_prefix(unsigned user, bool meta);
+
+extern const struct input_bind_map input_config_bind_map[RARCH_BIND_LIST_END_NULL];
 
 RETRO_END_DECLS
