+++ retroarch.c
@@ -25762,81 +25762,6 @@ void input_keyboard_event(bool down, unsigned code,
    }
 }
 
-static bool input_config_bind_map_get_valid(unsigned bind_index)
-{
-   const struct input_bind_map *keybind =
-      (const struct input_bind_map*)INPUT_CONFIG_BIND_MAP_GET(bind_index);
-   if (!keybind)
-      return false;
-   return keybind->valid;
-}
-
-unsigned input_config_bind_map_get_meta(unsigned bind_index)
-{
-   const struct input_bind_map *keybind =
-      (const struct input_bind_map*)INPUT_CONFIG_BIND_MAP_GET(bind_index);
-   if (!keybind)
-      return 0;
-   return keybind->meta;
-}
-
-const char *input_config_bind_map_get_base(unsigned bind_index)
-{
-   const struct input_bind_map *keybind =
-      (const struct input_bind_map*)INPUT_CONFIG_BIND_MAP_GET(bind_index);
-   if (!keybind)
-      return NULL;
-   return keybind->base;
-}
-
-const char *input_config_bind_map_get_desc(unsigned bind_index)
-{
-   const struct input_bind_map *keybind =
-      (const struct input_bind_map*)INPUT_CONFIG_BIND_MAP_GET(bind_index);
-   if (!keybind)
-      return NULL;
-   return msg_hash_to_str(keybind->desc);
-}
-
-uint8_t input_config_bind_map_get_retro_key(unsigned bind_index)
-{
-   const struct input_bind_map *keybind =
-      (const struct input_bind_map*)INPUT_CONFIG_BIND_MAP_GET(bind_index);
-   if (!keybind)
-      return 0;
-   return keybind->retro_key;
-}
-
-static const char *input_config_get_prefix(unsigned user, bool meta)
-{
-   static const char *bind_user_prefix[MAX_USERS] = {
-      "input_player1",
-      "input_player2",
-      "input_player3",
-      "input_player4",
-      "input_player5",
-      "input_player6",
-      "input_player7",
-      "input_player8",
-      "input_player9",
-      "input_player10",
-      "input_player11",
-      "input_player12",
-      "input_player13",
-      "input_player14",
-      "input_player15",
-      "input_player16",
-   };
-   if (meta)
-   {
-      if (user == 0)
-         return "input";
-      /* Don't bother with meta bind for anyone else than first user. */
-      return NULL;
-   }
-   return bind_user_prefix[user];
-}
-
 /**
  * input_config_translate_str_to_rk:
  * @str                            : String to translate to key ID.
@@ -25878,213 +25803,6 @@ unsigned input_config_translate_str_to_bind_id(const char *str)
          return i;
 
    return RARCH_BIND_LIST_END;
-}
-
-static uint16_t input_config_parse_hat(const char *dir)
-{
-   if (     dir[0] == 'u'
-         && dir[1] == 'p'
-         && dir[2] == '\0'
-      )
-      return HAT_UP_MASK;
-   else if (
-            dir[0] == 'd'
-         && dir[1] == 'o'
-         && dir[2] == 'w'
-         && dir[3] == 'n'
-         && dir[4] == '\0'
-         )
-      return HAT_DOWN_MASK;
-   else if (
-            dir[0] == 'l'
-         && dir[1] == 'e'
-         && dir[2] == 'f'
-         && dir[3] == 't'
-         && dir[4] == '\0'
-         )
-      return HAT_LEFT_MASK;
-   else if (
-            dir[0] == 'r'
-         && dir[1] == 'i'
-         && dir[2] == 'g'
-         && dir[3] == 'h'
-         && dir[4] == 't'
-         && dir[5] == '\0'
-         )
-      return HAT_RIGHT_MASK;
-
-   return 0;
-}
-
-static void input_config_parse_joy_button(
-      char *s,
-      config_file_t *conf, const char *prefix,
-      const char *btn, struct retro_keybind *bind)
-{
-   char tmp[64];
-   char key[64];
-   char key_label[64];
-   struct config_entry_list *tmp_a         = NULL;
-
-   tmp[0] = key[0] = key_label[0] = '\0';
-
-   fill_pathname_join_delim(key, s,
-         "btn", '_', sizeof(key));
-   fill_pathname_join_delim(key_label, s,
-         "btn_label", '_', sizeof(key_label));
-
-   if (config_get_array(conf, key, tmp, sizeof(tmp)))
-   {
-      btn = tmp;
-      if (     btn[0] == 'n'
-            && btn[1] == 'u'
-            && btn[2] == 'l'
-            && btn[3] == '\0'
-         )
-         bind->joykey = NO_BTN;
-      else
-      {
-         if (*btn == 'h')
-         {
-            const char *str = btn + 1;
-            /* Parse hat? */
-            if (str && ISDIGIT((int)*str))
-            {
-               char        *dir = NULL;
-               uint16_t     hat = strtoul(str, &dir, 0);
-               uint16_t hat_dir = dir ? input_config_parse_hat(dir) : 0;
-               if (hat_dir)
-                  bind->joykey = HAT_MAP(hat, hat_dir);
-            }
-         }
-         else
-            bind->joykey = strtoull(tmp, NULL, 0);
-      }
-   }
-
-   tmp_a = config_get_entry(conf, key_label);
-
-   if (tmp_a && !string_is_empty(tmp_a->value))
-   {
-      if (!string_is_empty(bind->joykey_label))
-         free(bind->joykey_label);
-
-      bind->joykey_label = strdup(tmp_a->value);
-   }
-}
-
-static void input_config_parse_joy_axis(
-      char *s,
-      config_file_t *conf, const char *prefix,
-      const char *axis, struct retro_keybind *bind)
-{
-   char       tmp[64];
-   char       key[64];
-   char key_label[64];
-   struct config_entry_list *tmp_a         = NULL;
-
-   tmp[0] = key[0] = key_label[0] = '\0';
-
-   fill_pathname_join_delim(key, s,
-         "axis", '_', sizeof(key));
-   fill_pathname_join_delim(key_label, s,
-         "axis_label", '_', sizeof(key_label));
-
-   if (config_get_array(conf, key, tmp, sizeof(tmp)))
-   {
-      if (     tmp[0] == 'n'
-            && tmp[1] == 'u'
-            && tmp[2] == 'l'
-            && tmp[3] == '\0'
-         )
-         bind->joyaxis = AXIS_NONE;
-      else if (strlen(tmp) >= 2 && (*tmp == '+' || *tmp == '-'))
-      {
-         int i_axis = (int)strtol(tmp + 1, NULL, 0);
-         if (*tmp == '+')
-            bind->joyaxis = AXIS_POS(i_axis);
-         else
-            bind->joyaxis = AXIS_NEG(i_axis);
-      }
-
-      /* Ensure that D-pad emulation doesn't screw this over. */
-      bind->orig_joyaxis = bind->joyaxis;
-   }
-
-   tmp_a = config_get_entry(conf, key_label);
-
-   if (tmp_a && (!string_is_empty(tmp_a->value)))
-   {
-      if (bind->joyaxis_label &&
-            !string_is_empty(bind->joyaxis_label))
-         free(bind->joyaxis_label);
-      bind->joyaxis_label = strdup(tmp_a->value);
-   }
-}
-
-static void input_config_parse_mouse_button(
-      char *s,
-      config_file_t *conf, const char *prefix,
-      const char *btn, struct retro_keybind *bind)
-{
-   int val;
-   char tmp[64];
-   char key[64];
-
-   tmp[0] = key[0] = '\0';
-
-   fill_pathname_join_delim(key, s, "mbtn", '_', sizeof(key));
-
-   if (config_get_array(conf, key, tmp, sizeof(tmp)))
-   {
-      bind->mbutton = NO_BTN;
-
-      if (tmp[0]=='w')
-      {
-         switch (tmp[1])
-         {
-            case 'u':
-               bind->mbutton = RETRO_DEVICE_ID_MOUSE_WHEELUP;
-               break;
-            case 'd':
-               bind->mbutton = RETRO_DEVICE_ID_MOUSE_WHEELDOWN;
-               break;
-            case 'h':
-               switch (tmp[2])
-               {
-                  case 'u':
-                     bind->mbutton = RETRO_DEVICE_ID_MOUSE_HORIZ_WHEELUP;
-                     break;
-                  case 'd':
-                     bind->mbutton = RETRO_DEVICE_ID_MOUSE_HORIZ_WHEELDOWN;
-                     break;
-               }
-               break;
-         }
-      }
-      else
-      {
-         val = atoi(tmp);
-         switch (val)
-         {
-            case 1:
-               bind->mbutton = RETRO_DEVICE_ID_MOUSE_LEFT;
-               break;
-            case 2:
-               bind->mbutton = RETRO_DEVICE_ID_MOUSE_RIGHT;
-               break;
-            case 3:
-               bind->mbutton = RETRO_DEVICE_ID_MOUSE_MIDDLE;
-               break;
-            case 4:
-               bind->mbutton = RETRO_DEVICE_ID_MOUSE_BUTTON_4;
-               break;
-            case 5:
-               bind->mbutton = RETRO_DEVICE_ID_MOUSE_BUTTON_5;
-               break;
-         }
-      }
-   }
 }
 
 static void input_config_get_bind_string_joykey(
@@ -26596,32 +26314,6 @@ const struct retro_keybind *input_config_get_bind_auto(
    return NULL;
 }
 
-void input_config_reset_autoconfig_binds(unsigned port)
-{
-   unsigned i;
-
-   if (port >= MAX_USERS)
-      return;
-
-   for (i = 0; i < RARCH_BIND_LIST_END; i++)
-   {
-      input_autoconf_binds[port][i].joykey  = NO_BTN;
-      input_autoconf_binds[port][i].joyaxis = AXIS_NONE;
-
-      if (input_autoconf_binds[port][i].joykey_label)
-      {
-         free(input_autoconf_binds[port][i].joykey_label);
-         input_autoconf_binds[port][i].joykey_label = NULL;
-      }
-
-      if (input_autoconf_binds[port][i].joyaxis_label)
-      {
-         free(input_autoconf_binds[port][i].joyaxis_label);
-         input_autoconf_binds[port][i].joyaxis_label = NULL;
-      }
-   }
-}
-
 void input_config_reset(void)
 {
    unsigned i;
@@ -26713,243 +26405,6 @@ void config_read_keybinds_conf(void *data)
          input_config_parse_mouse_button(str, conf, prefix, btn, bind);
       }
    }
-}
-
-void input_config_set_autoconfig_binds(unsigned port, void *data)
-{
-   unsigned i;
-   config_file_t *config       = (config_file_t*)data;
-   struct retro_keybind *binds = NULL;
-
-   if ((port >= MAX_USERS) || !config)
-      return;
-
-   binds = input_autoconf_binds[port];
-
-   for (i = 0; i < RARCH_BIND_LIST_END; i++)
-   {
-      const struct input_bind_map *keybind =
-         (const struct input_bind_map*)INPUT_CONFIG_BIND_MAP_GET(i);
-      if (keybind)
-      {
-         char str[256];
-         const char *base = keybind->base;
-         str[0]                     = '\0';
-
-         fill_pathname_join_delim(str, "input", base,  '_', sizeof(str));
-
-         input_config_parse_joy_button(str, config, "input", base, &binds[i]);
-         input_config_parse_joy_axis  (str, config, "input", base, &binds[i]);
-      }
-   }
-}
-
-/**
- * input_config_save_keybinds_user:
- * @conf               : pointer to config file object
- * @user               : user number
- *
- * Save the current keybinds of a user (@user) to the config file (@conf).
- */
-void input_config_save_keybinds_user(void *data, unsigned user)
-{
-   unsigned i = 0;
-   config_file_t *conf = (config_file_t*)data;
-
-   for (i = 0; input_config_bind_map_get_valid(i); i++)
-   {
-      char key[64];
-      char btn[64];
-      const struct input_bind_map *keybind =
-         (const struct input_bind_map*)INPUT_CONFIG_BIND_MAP_GET(i);
-      bool meta                            = keybind ? keybind->meta : false;
-      const char *prefix                   = input_config_get_prefix(user, meta);
-      const struct retro_keybind *bind     = &input_config_binds[user][i];
-      const char                 *base     = NULL;
-
-      if (!prefix || !bind->valid || !keybind)
-         continue;
-
-      base                                 = keybind->base;
-      key[0] = btn[0]                      = '\0';
-
-      fill_pathname_join_delim(key, prefix, base, '_', sizeof(key));
-
-      input_keymaps_translate_rk_to_str(bind->key, btn, sizeof(btn));
-      config_set_string(conf, key, btn);
-
-      input_config_save_keybind(conf, prefix, base, bind, true);
-   }
-}
-
-static void save_keybind_hat(config_file_t *conf, const char *key,
-      const struct retro_keybind *bind)
-{
-   char config[16];
-   unsigned hat     = (unsigned)GET_HAT(bind->joykey);
-   const char *dir  = NULL;
-
-   config[0]        = '\0';
-
-   switch (GET_HAT_DIR(bind->joykey))
-   {
-      case HAT_UP_MASK:
-         dir = "up";
-         break;
-
-      case HAT_DOWN_MASK:
-         dir = "down";
-         break;
-
-      case HAT_LEFT_MASK:
-         dir = "left";
-         break;
-
-      case HAT_RIGHT_MASK:
-         dir = "right";
-         break;
-
-      default:
-         break;
-   }
-
-   snprintf(config, sizeof(config), "h%u%s", hat, dir);
-   config_set_string(conf, key, config);
-}
-
-static void save_keybind_joykey(config_file_t *conf,
-      const char *prefix,
-      const char *base,
-      const struct retro_keybind *bind, bool save_empty)
-{
-   char key[64];
-
-   key[0] = '\0';
-
-   fill_pathname_join_delim_concat(key, prefix,
-         base, '_', "_btn", sizeof(key));
-
-   if (bind->joykey == NO_BTN)
-   {
-       if (save_empty)
-         config_set_string(conf, key, "nul");
-   }
-   else if (GET_HAT_DIR(bind->joykey))
-      save_keybind_hat(conf, key, bind);
-   else
-      config_set_uint64(conf, key, bind->joykey);
-}
-
-static void save_keybind_axis(config_file_t *conf,
-      const char *prefix,
-      const char *base,
-      const struct retro_keybind *bind, bool save_empty)
-{
-   char key[64];
-   unsigned axis   = 0;
-   char dir        = '\0';
-
-   key[0] = '\0';
-
-   fill_pathname_join_delim_concat(key,
-         prefix, base, '_',
-         "_axis",
-         sizeof(key));
-
-   if (bind->joyaxis == AXIS_NONE)
-   {
-      if (save_empty)
-         config_set_string(conf, key, "nul");
-   }
-   else if (AXIS_NEG_GET(bind->joyaxis) != AXIS_DIR_NONE)
-   {
-      dir = '-';
-      axis = AXIS_NEG_GET(bind->joyaxis);
-   }
-   else if (AXIS_POS_GET(bind->joyaxis) != AXIS_DIR_NONE)
-   {
-      dir = '+';
-      axis = AXIS_POS_GET(bind->joyaxis);
-   }
-
-   if (dir)
-   {
-      char config[16];
-
-      config[0] = '\0';
-
-      snprintf(config, sizeof(config), "%c%u", dir, axis);
-      config_set_string(conf, key, config);
-   }
-}
-
-static void save_keybind_mbutton(config_file_t *conf,
-      const char *prefix,
-      const char *base,
-      const struct retro_keybind *bind, bool save_empty)
-{
-   char key[64];
-
-   key[0] = '\0';
-
-   fill_pathname_join_delim_concat(key, prefix,
-      base, '_', "_mbtn", sizeof(key));
-
-   switch (bind->mbutton)
-   {
-      case RETRO_DEVICE_ID_MOUSE_LEFT:
-         config_set_uint64(conf, key, 1);
-         break;
-      case RETRO_DEVICE_ID_MOUSE_RIGHT:
-         config_set_uint64(conf, key, 2);
-         break;
-      case RETRO_DEVICE_ID_MOUSE_MIDDLE:
-         config_set_uint64(conf, key, 3);
-         break;
-      case RETRO_DEVICE_ID_MOUSE_BUTTON_4:
-         config_set_uint64(conf, key, 4);
-         break;
-      case RETRO_DEVICE_ID_MOUSE_BUTTON_5:
-         config_set_uint64(conf, key, 5);
-         break;
-      case RETRO_DEVICE_ID_MOUSE_WHEELUP:
-         config_set_string(conf, key, "wu");
-         break;
-      case RETRO_DEVICE_ID_MOUSE_WHEELDOWN:
-         config_set_string(conf, key, "wd");
-         break;
-      case RETRO_DEVICE_ID_MOUSE_HORIZ_WHEELUP:
-         config_set_string(conf, key, "whu");
-         break;
-      case RETRO_DEVICE_ID_MOUSE_HORIZ_WHEELDOWN:
-         config_set_string(conf, key, "whd");
-         break;
-      default:
-         if (save_empty)
-            config_set_string(conf, key, "nul");
-         break;
-   }
-}
-
-/**
- * input_config_save_keybind:
- * @conf               : pointer to config file object
- * @prefix             : prefix name of keybind
- * @base               : base name   of keybind
- * @bind               : pointer to key binding object
- * @kb                 : save keyboard binds
- *
- * Save a key binding to the config file.
- */
-void input_config_save_keybind(void *data, const char *prefix,
-      const char *base, const struct retro_keybind *bind,
-      bool save_empty)
-{
-   config_file_t *conf = (config_file_t*)data;
-
-   save_keybind_joykey (conf, prefix, base, bind, save_empty);
-   save_keybind_axis   (conf, prefix, base, bind, save_empty);
-   save_keybind_mbutton(conf, prefix, base, bind, save_empty);
 }
 
 /* MIDI */
